chore: extract last run common logic
