+++ web/app/components/workflow-app/hooks/index.ts
@@ -5,6 +5,6 @@ export * from './use-workflow-run'
 export * from './use-workflow-start-run'
 export * from './use-is-chat-mode'
 export * from './use-workflow-refresh-draft'
-export * from './use-fetch-workflow-inspect-vars'
+export * from '../../workflow/hooks/use-fetch-workflow-inspect-vars'
 export * from './use-inspect-vars-crud'
 export * from './use-configs-map'
